# openapi/SchedulingProviderV1Service.openapi3.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "SAP Advanced Financial Closing Scheduling Service Provider Interface",
    "description": "The Scheduling Service Provider Interface of SAP Advanced Financial Closing allows the integration of third-party scheduling systems with SAP Advanced Financial Closing. It includes the retrieval of job definitions, as well as the scheduling and synchronization of jobs. <a href=\"https://api.sap.com/api/SSPIV1/overview\" target=\"_blank\">See OpenAPI Specification on SAP Business Accelerator Hub</a>.",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "/api/job-scheduling/v1"
    }
  ],
...
  "paths": {
    "/Capabilities": {
      "get": {
        "summary": "Retrieves capabilities.",
        "tags": ["Capabilities"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Retrieved capabilities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sapafcsdk.scheduling.SchedulingProviderService.Capabilities"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "406": {
            "$ref": "#/components/responses/406"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "411": {
            "$ref": "#/components/responses/411"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "500": {
            "$ref": "#/components/responses/500"
          },
          "501": {
            "$ref": "#/components/responses/501"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "The operation allows you to retrieve capabilities."
      }
    },
    "/Job": {
      "get": {
        "summary": "Retrieves a list of job.",
        "tags": ["Job"],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "description": "The number of query results to skip. A value less than 0 SHALL be interpreted as 0. Example: ?skip=10",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "top",
            "in": "query",
            "required": false,
            "description": "Non-negative integer. Specifies the desired maximum number of query results to return per page. Example: ?top=10",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "referenceID",
            "in": "query",
            "required": false,
            "description": "Allows filtering of query results by referenceID. Example: ?referenceID=01234567-89ab-cdef-0123-456789abcdef",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "description": "Allows filtering of query results by name. Example: ?name=JOB_1 (optional)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved job",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "title": "Collection of Job",
                  "items": {
                    "$ref": "#/components/schemas/sapafcsdk.scheduling.SchedulingProviderService.Job"
                  }
                }
              }
            },
            "headers": {
              "x-total-count": {
                "description": "Total number of records.",
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "406": {
            "$ref": "#/components/responses/406"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "411": {
            "$ref": "#/components/responses/411"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "500": {
            "$ref": "#/components/responses/500"
          },
          "501": {
            "$ref": "#/components/responses/501"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "The operation allows you to retrieve a list of job based on the provided parameters."
      },
      "post": {
        "summary": "Creates a single job.",
        "tags": ["Job"],
        "requestBody": {
          "description": "New job",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/sapafcsdk.scheduling.SchedulingProviderService.Job-create"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sapafcsdk.scheduling.SchedulingProviderService.Job"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "406": {
            "$ref": "#/components/responses/406"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "411": {
            "$ref": "#/components/responses/411"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "500": {
            "$ref": "#/components/responses/500"
          },
          "501": {
            "$ref": "#/components/responses/501"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "The operation allows you to create a single job."
      }
    },
    "/Job/{ID}": {
      "parameters": [
        {
          "description": "key: ID",
          "in": "path",
          "name": "ID",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 255
          }
        }
      ],
      "get": {
        "summary": "Retrieves a single job.",
        "tags": ["Job"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Retrieved job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sapafcsdk.scheduling.SchedulingProviderService.Job"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "406": {
            "$ref": "#/components/responses/406"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "411": {
            "$ref": "#/components/responses/411"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "500": {
            "$ref": "#/components/responses/500"
          },
          "501": {
            "$ref": "#/components/responses/501"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "The operation allows you to retrieve a single job."
      }
    },
    "/Job/{ID}/parameters": {
      "parameters": [
        {
          "description": "key: ID",
          "in": "path",
          "name": "ID",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 255
          }
        }
      ],
      "get": {
        "summary": "Retrieves a list of parameters of a job.",
        "tags": ["Job"],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "description": "The number of query results to skip. A value less than 0 SHALL be interpreted as 0. Example: ?skip=10",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "top",
            "in": "query",
            "required": false,
            "description": "Non-negative integer. Specifies the desired maximum number of query results to return per page. Example: ?top=10",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "title": "Collection of JobParameter",
                  "items": {
                    "$ref": "#/components/schemas/sapafcsdk.scheduling.SchedulingProviderService.JobParameter"
                  }
                }
              }
            },
            "headers": {
              "x-total-count": {
                "description": "Total number of records.",
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "406": {
            "$ref": "#/components/responses/406"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "411": {
            "$ref": "#/components/responses/411"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "500": {
            "$ref": "#/components/responses/500"
          },
          "501": {
            "$ref": "#/components/responses/501"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "The operation allows you to retrieve a list of parameters of a job based on the provided parameters."
      }
    },
...
  },
  "components": {
    "schemas": {
      "sapafcsdk.scheduling.SchedulingProviderService.Capabilities": {
        "title": "Capabilities",
        "type": "object",
        "properties": {
          "supportsNotification": {
            "type": "boolean"
          },
          "applicationUrl": {
            "type": "string"
          }
        }
      },
...
    },
...
    "responses": {
      "400": {
        "description": "Incorrect request. Format or structure invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error"
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized. Action requires user authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error"
            }
          }
        }
      },
      "403": {
        "description": "Access forbidden. The caller does not have the required permissions to access the resource",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error"
            }
          }
        }
      },
      "404": {
        "description": "Resource not found. The caller either provided a wrong URL or the requested resource does not exist (any longer).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error"
            }
          }
        }
      },
...
      "409": {
        "description": "Request conflict. State of the resource does not permit the request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "description": "To access this API, use the OAuth 2.0 client credentials grant flow.",
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://{subdomain}.authentication.{region}.hana.ondemand.com/oauth/token",
            "scopes": {}
          }
        }
      },
      "bearer": {
        "description": "To access this API, use the HTTP bearer authentication.",
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "oauth2": []
    },
    {
      "bearer": []
    }
  ],
...
}